now when create an Order use batch verifyAndGetProducts

+++ order-service/src/main/java/com/ecommerce/orderservice/application/dto/OrderItemResponse.java
@@ -2,6 +2,8 @@ package com.ecommerce.orderservice.application.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
+import java.math.BigDecimal;
+
 @Schema(description = "Represents an item within an order")
 public record OrderItemResponse(
         @Schema(description = "Unique identifier of the product", example = "1")
@@ -11,5 +13,5 @@ public record OrderItemResponse(
         @Schema(description = "Quantity of the product ordered", example = "5")
         Integer quantity,
         @Schema(description = "Unit price of the product", example = "10.99")
-        Double unitPrice
+        BigDecimal unitPrice
 ) {}
